pyproject.toml
[tool.poetry]
name = "DevelopsTodayNewsBoard"
version = "0.1.0"
description = "DevelopsToday stage #2 test task"
authors = ["Ulukman Amangeldiev <ulukmanatageldiuulu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.8.6"
Pillow = "^8.2.0"
django-extensions = "^3.1.3"

[tool.poetry.dev-dependencies]
isort = "^5.8.0"
flake8 = "^3.9.2"
black = "^21.6b0"
drf-yasg = "^1.20.0"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
color = true
exclude = 'env/'
line_length = 79

[tool.isort]
skip_gitignore = true

[tool.autoflake]
remove_unused_variables = true
remove_all_unused_imports = true